requirements.txt
flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.1
flask-bcrypt==1.0.1
PyJWT==2.8.0
requests==2.31.0
python-dotenv==1.0.0
PyMuPDF==1.23.8
